third_party/WebKit/LayoutTests/fast/css-generated-content/details-before-after-content.html

<!DOCTYPE html>

<script src="../../resources/js-test.js"></script>

<style>
    details.before:before {
      content: "+";
      display: block;
    }
    details.after:after {
      content: "+";
      display: block;
    }
</style>

<details>
    <summary>Should allow generated content inside details</summary>
</details>

<script>
if (window.testRunner)
    testRunner.dumpAsText();

var details = document.querySelector('details');

function assertHeightChangedWithClassName(className)
{
    oldHeight = details.offsetHeight;
    details.className = className;
    newHeight = details.offsetHeight;
    debug('::' + className);
    shouldBeTrue("newHeight > oldHeight");
    details.className = '';
}

description("Should allow generated content inside details");
assertHeightChangedWithClassName('before');
assertHeightChangedWithClassName('after');
details.remove();
</script>
