third_party/blink/web_tests/fast/css-generated-content/details-before-after-content-expected.txt

Should allow generated content inside details

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

::before
PASS newHeight > oldHeight is true
::after
PASS newHeight > oldHeight is true
PASS successfullyParsed is true

TEST COMPLETE
